messaging_app/requirements.txt
asgiref==3.9.1
class-registry==2.1.2
Django==5.2.4
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
filters==1.3.2
mysqlclient==2.2.7
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
